backend/requirements.txt
# Core FastAPI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.1.0
slowapi>=0.1.9

# Anthropic SDK
anthropic>=0.72.0

# HTTP client
httpx>=0.25.1

# Environment and utils
python-dotenv>=1.0.0
python-multipart>=0.0.6
typer>=0.9.0
rich>=13.7.0

# Vector database
chromadb>=0.4.0

# Cohere for embeddings
cohere>=5.0.0

# Ollama
ollama>=0.4.0

# Browser automation
playwright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Data validation
marshmallow>=3.13.0,<3.23.0
environs>=9.5.0

# LangChain + LangGraph - Agent orchestration
langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-anthropic>=1.3.0
langchain-cohere>=0.4.0
langchain-ollama>=1.0.0
langchain-huggingface>=1.2.0
langgraph>=1.0.0

# LlamaIndex - RAG features
llama-index-core==0.11.23

# Web search
ddgs>=9.0.0

# Content extraction
trafilatura>=1.6.0

# Authentication
supabase>=2.0.0
PyJWT>=2.8.0
argon2-cffi>=23.1.0
email-validator>=2.0.0
